requirements.txt
pandas==2.2.2
Requests==2.32.3
streamlit==1.36.0
pillow==10.4.0